% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{add_filters}
\alias{add_filters}
\title{Add variables to a tplyr_meta object}
\usage{
add_filters(meta, filters)
}
\arguments{
\item{meta}{tplyr_meta object}

\item{...}{Variables to be added}
}
\value{
tplyr_meta object
}
\description{
Add variables to a tplyr_meta object
}
