% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_row_labels.R
\name{add_indentation}
\alias{add_indentation}
\title{Add indentation level based}
\usage{
add_indentation(.x, .y, indent = "  ")
}
\arguments{
\item{.x}{The number of levels to indent}

\item{.y}{Input variable for which indentation will be done}
}
\value{
Character string with indentation applied
}
\description{
Add indentation level based
}
