% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layering.R
\name{add_layer}
\alias{add_layer}
\title{Attach a layer to a \code{tplyr_table} object}
\usage{
add_layer(parent, layer)
}
\arguments{
\item{layer}{}
}
\value{

}
\description{
Attach a layer to a \code{tplyr_table} object
}
