% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{add_missing_subjects_row}
\alias{add_missing_subjects_row}
\title{Add a missing subject row into a count summary.}
\usage{
add_missing_subjects_row(e, fmt = NULL, sort_value = NULL)
}
\arguments{
\item{e}{A \code{count_layer} object}

\item{fmt}{An f_str object used to format the total row. If none is provided,
display is based on the layer formatting.}

\item{sort_value}{The value that will appear in the ordering column for total
rows. This must be a numeric value.}
}
\description{
Add a missing subject row into a count summary.
}
\examples{

tplyr_table(mtcars, gear) \%>\%
  add_layer(
    group_count(cyl) \%>\%
      add_missing_subjects_row(f_str("xxxx", n))
   ) \%>\%
   build()
}
