% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{add_variables}
\alias{add_variables}
\title{Add variables to a tplyr_meta object}
\usage{
add_variables(meta, names)
}
\arguments{
\item{meta}{tplyr_meta object}

\item{names}{Variables to be added}
}
\value{
tplyr_meta object
}
\description{
Add variables to a tplyr_meta object
}
