% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_has_class}
\alias{assert_has_class}
\title{Assert that an argument is of a certain class}
\usage{
assert_has_class(x, should_be)
}
\arguments{
\item{should_be}{}
}
\value{

}
\description{
Assert that an argument is of a certain class
}
