% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_inherits_class}
\alias{assert_inherits_class}
\title{Assert that an argument inherits certain class}
\usage{
assert_inherits_class(x, should_have)
}
\arguments{
\item{should_have}{}
}
\value{

}
\description{
Assert that an argument inherits certain class
}
