% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{construct_count_string}
\alias{construct_count_string}
\title{Format n counts for display in count_layer}
\usage{
construct_count_string(.n, .total, count_fmt = NULL)
}
\arguments{
\item{ns}{}
}
\value{
A tibble replacing the originial counts
}
\description{
Format n counts for display in count_layer
}
