% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{construct_desc_string}
\alias{construct_desc_string}
\title{Format a descriptive statistics display string}
\usage{
construct_desc_string(..., .fmt_str = NULL)
}
\arguments{
\item{...}{A row of a data frame - captured through the ellipsis argument}

\item{.fmt_str}{The format strings container with all f_str objects and row labels}
}
\value{
A character vector of display formatted strings
}
\description{
Intended to be applied through a map_chr call
}
