% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{depth_from_table}
\alias{depth_from_table}
\title{Find depth of a layer object}
\usage{
depth_from_table(layer, i)
}
\arguments{
\item{layer}{A layer object}

\item{i}{The current index}
}
\value{
the number of containers a layer is in
}
\description{
This function returns the number of containers "above" a layer object. As
layers can be nested layers may contain layers and so on. This uses
recursion to find the table environment
}
