% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{dmessage}
\alias{dmessage}
\title{Title}
\usage{
dmessage(message, ...)
}
\arguments{
\item{message}{The debugging message to be printed}

\item{...}{Additional parameters passed to \code{rlang::inform}}
}
\value{
Nothing
}
\description{
Title
}
\examples{
tplyr_debug(TRUE)
dmessage('This is will print!')

tplyr_debug(FALSE)
dmessage('This will not...')
}
