% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{extract_character_from_quo}
\alias{extract_character_from_quo}
\title{Take a list of quosures and pull out things that aren't symbols}
\usage{
extract_character_from_quo(var_list)
}
\arguments{
\item{var_list}{List of quosures}
}
\value{
Quosures that aren't symbols
}
\description{
Take a list of quosures and pull out things that aren't symbols
}
