% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{f_str}
\alias{f_str}
\title{Create a \code{f_str} object}
\usage{
f_str(format_string, ..., empty = "")
}
\arguments{
\item{format_string}{The desired display format. X's indicate digits. On the left, the number of x's indicates the integer length. On the
right, the number of x's controls decimal precision and rounding. Variables are inferred by any separation of the 'x' values other than a
decimal.}

\item{...}{The variables to be formatted using the format specified in \code{format_string}.}

\item{empty}{The string to display when the numeric data is not available}
}
\value{
A \code{f_str} object, built on a list with two elements:
\describe{
\item{\code{format_string}}{The specified format string for display}
\item{\code{vars}}{A list of names containing the variables that will be used to created the formatted display string}
}
}
\description{
The \code{f_str} object carries information that powers a significant amount of layer processing. The \code{format_string} parameter is
capable of controlling display of a data point and decimal precision. The variables provided in \code{...} control the values from the
data the output a particular formatted display value.
}
\examples{
f_str("xx.x (xx.x)", mean, sd)

}
