% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{fetch_formatting_info}
\alias{fetch_formatting_info}
\title{Placeholder function to fetch table formatting data from layers}
\usage{
fetch_formatting_info(x)
}
\arguments{
\item{x}{A tplyr_table object}
}
\value{
The data used to format layers. Structure currently TBD
}
\description{
Placeholder function to fetch table formatting data from layers
}
