% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_bindings.R
\name{get_layer_formatter}
\alias{get_layer_formatter}
\alias{set_layer_formatter}
\title{Set or return layer formatter}
\usage{
get_layer_formatter(layer)

set_layer_formatter(layer, formatter)
}
\arguments{
\item{layer}{A \code{tplyr_layer} object}

\item{formatter}{A function used to create the string formats for the
resulting numbers in output presentation.}
}
\value{
For \code{layer_formatter}, the formetter function bound to the
  supplied layer. For \code{set_layer_formatter} the modified layer environment.
}
\description{
Set or return layer formatter
}
\examples{
iris$Species2 <- iris$Species
lay <- tplyr_table(iris, Species) \%>\%
  group_count(Species) \%>\%
  set_layer_formatter(as.numeric)
}
