% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{gather_settings}
\alias{gather_settings}
\title{Gather the settings for a specific format string section}
\usage{
gather_settings(x)
}
\arguments{
\item{x}{A character string representing a format string section}
}
\value{
A named list of settings
}
\description{
This function will collect specific settings about a format string section,
including integer and decimal length, whether autos were turned on, and hug
character settings/
}
