% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_bindings.R
\name{get_custom_summaries}
\alias{get_custom_summaries}
\title{Get custom user summaries added to the layer environment via \code{set_custom_summaries}}
\usage{
get_custom_summaries(e = caller_env())
}
\arguments{
\item{e}{Environment to extract custom summaries from}
}
\value{
\code{custom_summaries} binding in the layer environment
}
\description{
This method is an internal intended to be called within the \code{tplyr_layer} environment
}
