% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data_labels}
\alias{get_data_labels}
\title{Get Data Labels}
\usage{
get_data_labels(data)
}
\arguments{
\item{data}{A Tplyr data set.}
}
\value{
A data.frame with columns `name` and `label` containing the names and labels of each column.
}
\description{
Get labels for data sets included in Tplyr.
}
