% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denom.R
\name{get_denom_total}
\alias{get_denom_total}
\title{Get the denominator used in shift pcts}
\usage{
get_denom_total(.data, denoms_by, denoms_df)
}
\arguments{
\item{.data}{A data.frame that has been grouped}

\item{denoms_by}{The variables used to get the denoms from}

\item{denoms_df}{The denoms_df that is created during layer processing.
Contains the unique combinations of all layer parameters and their counts.}
}
\value{
A data.frame with the
}
\description{
This is meant to be called in a dplyr grouped context
}
