% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denom.R
\name{get_header_n_value}
\alias{get_header_n_value}
\title{Get the header_n value of a certain}
\usage{
get_header_n_value(x, ...)
}
\arguments{
\item{x}{A tplyr_table or data.frame object}

\item{...}{Named parameters to filter the header_n data.frame}
}
\value{
The numeric value in
}
\description{
Get the header_n value of a certain
}
\examples{
library(dplyr)

t <- tplyr_table(mtcars, gear, cols = vars(cyl, am))

get_header_n_value(t, 3, 6, 0)
# Returns the number of cars that have 3 gears, 6 cyinders, and auto trasmission
}
