% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{get_meta_result}
\alias{get_meta_result}
\title{Extract the result metadata of a Tplyr table}
\usage{
get_meta_result(t, row_id, column)
}
\arguments{
\item{t}{A built Tplyr table}

\item{row_id}{The row_id value of the desired cell}

\item{column}{The result column of interest}
}
\value{
a tplyr_meta object
}
\description{
Given a row_id value and a result column, this function will return the
tplyr_meta object associated with that 'cell'.
}
\examples{

# TODO
}
