% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{get_meta_subset}
\alias{get_meta_subset}
\title{Extra a subset of the target data of a Tplyr table}
\usage{
get_meta_subset(t, row_id, column, add_cols = vars(USUBJID))
}
\arguments{
\item{t}{A built Tplyr table}

\item{row_id}{The row_id value of the desired cell}

\item{column}{The result column of interest}

\item{add_cols}{Additional columns}
}
\value{

}
\description{
Given a row_id value and a result column, this function will return the
subset of data from the target dataset which was used to produce that "cell"
}
\examples{

# TODO
}
