% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denom.R
\name{get_shift_total}
\alias{get_shift_total}
\title{Get the denominator used in shift pcts}
\usage{
get_shift_total(.data, denom_by, denoms_df)
}
\arguments{
\item{.data}{A data.frame that has been grouped}

\item{denom_by}{The variables}
}
\value{
A data.frame with the
}
\description{
This is meant to be called in a dplyr grouped context
}
