% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{get_summaries}
\alias{get_summaries}
\title{Get the summaries to be passed forward into \code{dplyr::summarize()}}
\usage{
get_summaries(e = caller_env())
}
\arguments{
\item{e}{the environment summaries are stored in.}
}
\value{
A list of expressions to be unpacked in \code{dplyr::summarize}
}
\description{
Get the summaries to be passed forward into \code{dplyr::summarize()}
}
