% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_target_levels}
\alias{get_target_levels}
\title{Get the unique levels/factors of a dataset}
\usage{
get_target_levels(e, x)
}
\arguments{
\item{e}{An environment, generally a table or a layer object}

\item{target_var}{A target variable to get the levels/unique values of}
}
\value{
Unique target values
}
\description{
Get the unique levels/factors of a dataset
}
