% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layering.R
\name{group_count}
\alias{group_count}
\title{Create a \code{count} layer for discrete count based summaries}
\usage{
group_count(parent, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Create a \code{count} layer for discrete count based summaries
}
