% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layering.R
\name{group_desc}
\alias{group_desc}
\title{Create a \code{desc} layer for descriptive statistics summaries}
\usage{
group_desc(parent, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Create a \code{desc} layer for descriptive statistics summaries
}
