% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layering.R
\name{group_shift}
\alias{group_shift}
\title{Create a \code{count} layer for shift count summaries}
\usage{
group_shift(parent, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Create a \code{count} layer for shift count summaries
}
