% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{has_format_strings}
\alias{has_format_strings}
\title{Check if format strings have been applied to a layer}
\usage{
has_format_strings(e)
}
\arguments{
\item{e}{Layer environment}
}
\value{
Boolean
 @noRd
}
\description{
Check if format strings have been applied to a layer
}
