% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_headers.R
\name{has_spanners}
\alias{has_spanners}
\title{Check if spanners exist}
\usage{
has_spanners(string_segments)
}
\arguments{
\item{string_segments}{Processed string segments from \code{identify_string_segments}}
}
\value{
Boolean
}
\description{
Check if spanners exist
}
