% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_bindings.R
\name{tplyr_header_n}
\alias{tplyr_header_n}
\alias{tplyr_header_n<-}
\alias{set_tplyr_header_n}
\title{Return or set header_n binding}
\usage{
tplyr_header_n(table)

tplyr_header_n(x) <- value

set_tplyr_header_n(table, header_n)
}
\arguments{
\item{table}{A \code{tplyr_table} object}

\item{x}{A \code{tplyr_table} object}

\item{value}{A named numeric vector. Names of vector should match headers.}

\item{header_n}{A named numeric vector. Names of vector should match headers.}
}
\value{
For \code{tplyr_header_n} the header_n binding of the
  \code{tplyr_table} object. For \code{tplyr_header_n<-} and
  \code{set_tplyr_header_n} the modified object.
}
\description{
Return or set header_n binding
}
