% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_bindings.R
\name{tplyr_header}
\alias{tplyr_header}
\alias{tplyr_header<-}
\alias{set_tplyr_header}
\title{Return or set header bindings}
\usage{
tplyr_header(table)

tplyr_header(x) <- value

set_tplyr_header(table, headers)
}
\arguments{
\item{table}{A \code{tplyr_table} object.}

\item{x}{A \code{tplyr_table} object}

\item{value}{A character vector detailing the column headers}

\item{headers}{A character vector detailing the column headers}
}
\value{
For \code{tplyr_header} the header binding of the \code{tplyr_talbe}
  object. For \code{tplyr_header<-} and \code{set_tplyr_header} the modified
  object.
}
\description{
The column header that will be displayed when the table is rendered.
}
\examples{
tab <- tplyr_table(iris, Species)

tplyr_header(tab) <- c("Sepal.Length", "Sepal.Width", "Petal.Length",
                       "Petal.Width", "Species")

}
