% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_exact}
\alias{match_exact}
\title{Convert a list of quosures to character strings}
\usage{
match_exact(var_list)
}
\arguments{
\item{vars}{List of quosures containing variables}
}
\value{
Character string of labels
}
\description{
Intended for use in a tidyselect context. Pivots take arguments as character strings or indices. Tidyselect tools return those
indices. This allows you to pass a list of quosures (which Tplyr carries a lot of) without explicitly converting types
}
\examples{
iris \%>\%
  group_by(Species) \%>\%
  summarize(mean=mean(Sepal.Length), median = median(Sepal.Length)) \%>\%
  pivot_longer(cols = match_exact(vars(mean, median)))

}
