% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modify_nested_call}
\alias{modify_nested_call}
\title{Extract the top function from a nested call and insert desired arguments}
\usage{
modify_nested_call(c, ...)
}
\arguments{
\item{c}{An R expression}

\item{...}{Parameters to insert into topmost call}
}
\value{
The original call object with
}
\description{
Extract the top function from a nested call and insert desired arguments
}
