% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{name_translator}
\alias{name_translator}
\title{Extract a translation vector for f_str objects}
\usage{
name_translator(fmt_strings)
}
\arguments{
\item{fmt_strings}{The \code{format_strings} varaible in a layer}
}
\value{
A named character vector with the flipping applied
}
\description{
The names of the format_strings list should be row labels in the output. The first
element of the \code{vars} object are the transpose variables, so make the names of
those variables the vector names, and the names of the format_strings elements the
values to allow easy creation of a \code{row_labels} variable in the data
}
