% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{num_fmt}
\alias{num_fmt}
\title{Format a numeric value using an \code{f_str} object}
\usage{
num_fmt(val, i, fmt = NULL, autos = NULL)
}
\arguments{
\item{val}{Numeric value to be formatted}

\item{i}{Index of the format within the \code{f_str} object}

\item{fmt}{\code{f_str} object with formatting information related to numeric value to be formatted}
}
\value{
String formatted numeric value
}
\description{
Using the \code{f_str} object, information about the integer length and
significant digits are extracted. Proper round is applied and the formatted numeric value is returned.
}
