% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tplyr_table}
\alias{print.tplyr_table}
\title{Print tplyr_table}
\usage{
\method{print}{tplyr_table}(x, ...)
}
\arguments{
\item{x}{A tplyr_table to display}

\item{...}{A paranthesis passed through from print dispatch

 @export
 @noRd}
}
\description{
Print tplyr_table
}
