% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{process_count_layer}
\alias{process_count_layer}
\title{Process a layer of type \code{group_count}}
\usage{
process_count_layer(e)
}
\arguments{
\item{e}{A layer environment}
}
\value{
A
}
\description{
Process a layer of type \code{group_count}
}
