% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{process_desc_layer}
\alias{process_desc_layer}
\title{Process a layer of type \code{desc}}
\usage{
process_desc_layer(e, summaries)
}
\arguments{
\item{e}{Environment within which a}
}
\value{
Processed data within layer environment
}
\description{
Process a layer of type \code{desc}
}
