% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{process_formatting}
\alias{process_formatting}
\title{Process layers to get formatted and pivoted tables.}
\usage{
process_formatting(x, ...)
}
\arguments{
\item{...}{arguments passed to dispatch}

\item{lay}{A tplyr_layer object}
}
\value{
The formatted_table object that is binded to the layer
}
\description{
This is an internal method, but is exported to support S3 dispatch. Not intended for direct use by a user.
}
