% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{process_metadata}
\alias{process_metadata}
\title{Process layers to get metadata tables}
\usage{
process_metadata(x, ...)
}
\arguments{
\item{x}{A tplyr_layer object}

\item{...}{arguments passed to dispatch}
}
\value{
The formatted_meta object that is bound to the layer
}
\description{
This is an internal method, but is exported to support S3 dispatch. Not intended for direct use by a user.
}
\keyword{internal}
