% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{process_summaries.desc_layer}
\alias{process_summaries.desc_layer}
\title{Process numeric data for a layer of type \code{desc}}
\usage{
\method{process_summaries}{desc_layer}(x, ...)
}
\arguments{
\item{x}{Environment within which a layer is processed}
}
\value{
Processed data within layer environment
}
\description{
Process numeric data for a layer of type \code{desc}
}
