% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{quo_class}
\alias{quo_class}
\title{Return the class of the expression inside a quosure}
\usage{
quo_class(q)
}
\arguments{
\item{q}{A quosure}
}
\value{
The class of quosure \code{q}
}
\description{
Return the class of the expression inside a quosure
}
\examples{
library(rlang)

q <- quo(a)
quo_class(q)

q <- quo("Hello!")
quo_class(q)

q <- quo(x + y)
quo_class(q)
}
