% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_by_string_names}
\alias{replace_by_string_names}
\title{Organize row labels within a layer output}
\usage{
replace_by_string_names(dat, by)
}
\arguments{
\item{dat}{A data.frame/tibble to have row labels renamed}

\item{by}{The \code{by} object within a layer}
}
\value{
A tibble with renamed variables and row labels re-ordered to the front of the tibble
}
\description{
Organize row labels within a layer output
}
