% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_leading_whitespace.R
\name{replace_leading_whitespace}
\alias{replace_leading_whitespace}
\title{Reformat strings with leading whitespace for HTML}
\usage{
replace_leading_whitespace(x, tab_width = 4)
}
\arguments{
\item{x}{Target string}

\item{tab_width}{Number of spaces to compensate for tabs}
}
\value{
String with &nbsp; replaced for leading whitespace
}
\description{
Reformat strings with leading whitespace for HTML
}
\examples{
x <- c(" Hello there", "  Goodbye Friend ",  "\tNice to meet you",
"  \t What are you up to? \t \t ")
replace_leading_whitespace(x)

replace_leading_whitespace(x, tab=2)

}
