% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskdiff.R
\name{riskdiff}
\alias{riskdiff}
\title{Calculate risk difference}
\usage{
riskdiff(diff_group, n_comp, n_ref, total_comp, total_ref, args = list(), ...)
}
\arguments{
\item{diff_group}{The concateations of the reference and comparison groups}

\item{n_comp}{The count of of the comparison}

\item{n_ref}{The count of the reference}

\item{total_comp}{The total of the comp}

\item{total_ref}{The total of the reference}

\item{args}{Arguments that will be passed into prop.test, provided as a named list}

\item{...}{}
}
\value{
A dataframe containing the group, the proportions of each comparator, the difference,
and the lower and upper CI
}
\description{
This function is intended to be called using \code{pmap_dfr} against the prepared two-day data
from \code{prep_two_way}
}
