% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_byrow_numeric_value}
\alias{set_byrow_numeric_value}
\title{Set the value the count layer is sorting on when the method is byrow.}
\usage{
set_byrow_numeric_value(e, byrow_numeric_value)
}
\arguments{
\item{e}{A \code{group_count} layer}

\item{byrow_numeric_value}{The numeric value the ordering will be done on.
This can be either n, distinct_n, pct, or distinct_pct. Due to the
evaluation of the layer you can add a value that isn't actually being
evaluated, if this happens this will only error out in the ordering.}
}
\value{
The modified layer object
}
\description{
The sorting logic will pull out the numeric value in the count layer and
add that as an ordering column. You can select which numeric value to pull
from.
}
