% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_count_fmt}
\alias{set_count_fmt}
\title{Set Count Layer String Format}
\usage{
set_count_fmt(x, str)
}
\arguments{
\item{x}{the layer object to add/modify the count format}

\item{str}{The f_str object to add}
}
\value{

}
\description{
Set Count Layer String Format
}
