% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_count_row_prefix}
\alias{set_count_row_prefix}
\title{Set the option to prefix the row_label values when a layer is formatted}
\usage{
set_count_row_prefix(e, count_row_prefix)
}
\arguments{
\item{e}{A tplyr_count layer}

\item{count_row_prefix}{A character to prefix the row labels}
}
\value{
The modified count_layer environment
}
\description{
This is generally used internally with a nested count layer.
}
\examples{
# Load in pipe
library(magrittr)

tplyr_table(mtcars, gear) \%>\%
  add_layer(
    group_count(cyl) \%>\%
      set_count_row_prefix("\t")
  ) \%>\%
  build()

}
