% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_denom_ignore}
\alias{set_denom_ignore}
\title{Set values the denominator calcuation will ignore}
\usage{
set_denom_ignore(e, ...)
}
\arguments{
\item{e}{A count_layer object}

\item{...}{Quoted values.}
}
\value{
The modified layer object
}
\description{
This is generally used for missing values. Values like "", NA, "NA" are common.
}
