% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_distinct_by}
\alias{set_distinct_by}
\title{Set counts to be distinct by some grouping variable.}
\usage{
set_distinct_by(e, distinct_by)
}
\arguments{
\item{e}{A count_layer object}

\item{distinct_by}{A variable to get the distinct}
}
\value{
The layer object with
}
\description{
Occasionally summaries call for counting distint values within a group. \code{set_distinct_by}
will update a count layer to only count distinct values by the specified variables.
}
\examples{
#Load in pipe
library(magrittr)

tplyr_table(mtcars, gear) \%>\%
  add_layer(
    group_count(cyl) \%>\%
      set_distinct_by(carb)
  ) \%>\%
  build()
}
