% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{set_format_strings.count_layer}
\alias{set_format_strings.count_layer}
\title{Set Count Layer String Format}
\usage{
\method{set_format_strings}{count_layer}(e, ...)
}
\arguments{
\item{e}{Layer on which to bind format strings}

\item{...}{Named parmeters containing calls to \code{f_str} to set the format strings}
}
\value{
Returns the modified layer object.
}
\description{
Set Count Layer String Format
}
\examples{
# TBD
}
