% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_include_total_row}
\alias{set_include_total_row}
\title{Set the include_total_row option for count processing}
\usage{
set_include_total_row(x, include_total)
}
\arguments{
\item{x}{A layer object}

\item{include_total}{A logical vector}
}
\description{
Set the include_total_row option for count processing
}
