% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_indentation}
\alias{set_indentation}
\title{Set the option to prefix the row_labels in the inner count_layer}
\usage{
set_indentation(e, indentation)
}
\arguments{
\item{e}{A \code{count_layer} object}

\item{indentation}{A character to prefix the row labels in an inner
count layer}
}
\value{
The modified count_layer environment
}
\description{
When a count layer uses nesting (i.e. triggered by \code{\link{set_nest_count}}),
the \code{indentation} argument's value will be used as a prefix for the inner layer's
records
}
