% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_inner_count_layer_prefix}
\alias{set_inner_count_layer_prefix}
\title{Set the option to prefix the row_labels in the inner count_layer}
\usage{
set_inner_count_layer_prefix(e, inner_count_layer_prefix)
}
\arguments{
\item{e}{A tplyr_count layer}

\item{inner_count_layer_prefix}{A character to prefix the row labels in an inner
count layer}
}
\value{
The modified count_layer environment
}
\description{
This is generally used internally with a nested count layer.
}
