% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_limit_data_by.R
\name{set_limit_data_by}
\alias{set_limit_data_by}
\title{Set variables to limit to data values only rather than fully completing all
possible levels}
\usage{
set_limit_data_by(e, ...)
}
\arguments{
\item{e}{A tplyr_layer}

\item{...}{Subset of variables within by or target variables}
}
\description{
Set variables to limit to data values only rather than fully completing all
possible levels
}
