% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_missing_subjects_row_label}
\alias{set_missing_subjects_row_label}
\title{Set the label for the missing subjects row}
\usage{
set_missing_subjects_row_label(e, missing_subjects_row_label)
}
\arguments{
\item{e}{A \code{count_layer} object}

\item{missing_subjects_row_label}{A character to label the total row}
}
\value{
The modified \code{count_layer} object
}
\description{
Set the label for the missing subjects row
}
\examples{

t <- tplyr_table(mtcars, gear) \%>\%
  add_layer(
    group_count(cyl) \%>\%
      add_missing_subjects_row() \%>\%
      set_missing_subjects_row_label("Missing")
  )
build(t)
}
