% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_nest_count}
\alias{set_nest_count}
\title{Set the option to nest count layers}
\usage{
set_nest_count(e, nest_count)
}
\arguments{
\item{e}{A tplyr_count layer}

\item{nest_count}{A logical value to set the nest option}
}
\value{
The modified layer
}
\description{
Set the option to nest count layers
}
