% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_numeric_where}
\alias{set_numeric_where}
\title{Title}
\usage{
set_numeric_where(e, numeric_cutoff, stat, column = NULL)
}
\arguments{
\item{e}{A \code{count_layer} object}

\item{numeric_cutoff}{A numeric value where only values greater than or equal
to will be displayed.}

\item{stat}{The statistic to use when filtering out rows. Either 'n',
'distinct_n', or 'pct' are allowable}

\item{column}{If only a particular column should be used to cutoff values, it
can be supplied here as a character value.}
}
\value{
The modified Tplyr layer object
}
\description{
Title
}
