% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_order_count_method}
\alias{set_order_count_method}
\title{Set the ordering logic for the count layer}
\usage{
set_order_count_method(e, order_count_method)
}
\arguments{
\item{e}{A \code{group_count} layer}

\item{order_count_method}{The logic determining how the rows in the final
layer output will be indexed. Options are 'byrow', 'byfactor', and
'byvarn'.}
}
\value{
The modifed layer object
}
\description{
Count layers are generally displayed in a descending order for a particular
treatment group. However Tplyr also supports displaying based on the factor
of the treatment variable, or on a specified <VAR>N variable already in the
dataset (i.e. VISIT <-> VISITN, TRT <-> TRTN). Defaults to 'byrow'.
}
