% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_ordering_cols}
\alias{set_ordering_cols}
\title{Set the columns to use when ordering the data in the count layer}
\usage{
set_ordering_cols(e, ...)
}
\arguments{
\item{e}{A tplyr \code{group_count} object}

\item{...}{Unquoted variables used to select the columns whose values will be extracted for ordering.}
}
\value{

}
\description{
Set the columns to use when ordering the data in the count layer
}
