% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_outer_inf}
\alias{set_outer_inf}
\title{Set the value of a outer nested count layer to Inf or -Inf}
\usage{
set_outer_inf(e, outer_inf)
}
\arguments{
\item{e}{A count layer}

\item{outer_inf}{A logical value. If the value is set to TRUE. The outer
count layer is set to Inf, if it is false it is set to -Inf.}
}
\value{
The modified count layer.
}
\description{
Set the value of a outer nested count layer to Inf or -Inf
}
