% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_outer_sort_position}
\alias{set_outer_sort_position}
\title{Set the value of a outer nested count layer to Inf or -Inf}
\usage{
set_outer_sort_position(e, outer_sort_position)
}
\arguments{
\item{e}{A count layer}

\item{outer_sort_position}{Either 'asc' or 'desc'. If desc the final ordering helper
will be set to Inf, if 'asc' the ordering helper is set to -Inf.}
}
\value{
The modified count layer.
}
\description{
Set the value of a outer nested count layer to Inf or -Inf
}
