% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_bindings.R
\name{layer_sort}
\alias{layer_sort}
\alias{set_layer_sort}
\title{Set or return sort layer binding}
\usage{
layer_sort(layer)

set_layer_sort(layer, sort)
}
\arguments{
\item{layer}{A \code{tplyr_layer} object}

\item{sort}{A string containing the sort method.}
}
\value{
For \code{layer_sort}, the sort binding of the layer object. For
  \code{set_layer_sort}, the modified layer environment.
}
\description{
Set or return sort layer binding
}
