% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_bindings.R
\name{sort_vars}
\alias{sort_vars}
\alias{set_sort_vars}
\title{Return or set sort_vars layer binding}
\usage{
sort_vars(layer)

set_sort_vars(layer, sort_vars)
}
\arguments{
\item{layer}{}

\item{sort_vars}{A character vector to sort the results of the summary.}
}
\value{
For \code{sort_vars}, the bindings of the layer object. For
  \code{set_sort_vars}, the modified layer environment.
}
\description{
Return or set sort_vars layer binding
}
