% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_bindings.R
\name{target_var}
\alias{target_var}
\alias{set_target_var}
\title{Set or return treat_var binding}
\usage{
target_var(layer)

set_target_var(layer, target_var)
}
\arguments{
\item{layer}{A \code{tplyr_layer} object}

\item{target_var}{A symbol to perform the analysis on}
}
\value{
For \code{treat_var}, the treatment variable binding of the layer
  object. For \code{set_treat_var}, the modified layer environment.
}
\description{
Set or return treat_var binding
}
