% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denom.R
\name{this_denom}
\alias{this_denom}
\title{Get the value of the header_n value in a grouped df}
\usage{
this_denom(.data, header_n)
}
\arguments{
\item{.data}{A tibble that has been grouped}

\item{header_n}{The header_n binding from a \code{tplyr_table} object.}
}
\value{
A single numeric value from the header_n binding that is pulled
 from the groups
}
\description{
This is intended to be called in a dplyr context using `do`. This function
returns the value of the header_n for the current group. `do` is currently
labeled as questioning but its replacement, group_map, is experimental
}
\examples{
library(magrittr)

t <- tplyr_table(mtcars, gear)

mtcars \%>\%
  group_by(gear, vs) \%>\%
  do(this_denom(., header_n(t)))

}
