% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tplyr_adae}
\alias{tplyr_adae}
\title{ADAE Data}
\format{
A data.frame with 276 rows and 55 columns.
}
\source{
https://github.com/phuse-org/TestDataFactory
}
\usage{
tplyr_adae
}
\description{
A subset of the PHUSE Test Data Factory ADAE data set.
}
\seealso{
[get_data_labels()]
}
\keyword{datasets}
