% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tplyr_adlb}
\alias{tplyr_adlb}
\title{ADLB Data}
\format{
A data.frame with 311 rows and 46 columns.
}
\source{
https://github.com/phuse-org/TestDataFactory
}
\usage{
tplyr_adlb
}
\description{
A subset of the PHUSE Test Data Factory ADLB data set.
}
\seealso{
[get_data_labels()]
}
\keyword{datasets}
