% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tplyr_adpe}
\alias{tplyr_adpe}
\title{ADPE Data}
\format{
A data.frame with 21 rows and 8 columns.
}
\usage{
tplyr_adpe
}
\description{
A mock-up dataset that is fit for testing data limiting
}
\keyword{datasets}
