% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tplyr_adsl}
\alias{tplyr_adsl}
\title{ADSL Data}
\format{
A data.frame with 254 rows and 49 columns.
}
\source{
https://github.com/phuse-org/TestDataFactory
}
\usage{
tplyr_adsl
}
\description{
A subset of the PHUSE Test Data Factory ADSL data set.
}
\seealso{
[get_data_labels()]
}
\keyword{datasets}
