% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{tplyr_debug}
\alias{tplyr_debug}
\title{Toggle 'Tplyr' debug messaging}
\usage{
tplyr_debug(status)
}
\arguments{
\item{status}{TRUE/FALSE - Turns status messaging on or off}
}
\value{
Nothing
}
\description{
\strong{NOTE:} This feature is being questioned and is not currently being utilized.
}
\examples{
tplyr_debug(TRUE)
}
