% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{tplyr_order}
\alias{tplyr_order}
\title{Set the sorting of rows/columns in a `tplyr_table` and `tplyr_layer` object}
\usage{
tplyr_order(x, ...)
}
\arguments{
\item{x}{A \code{data.frame} or \code{tplyr_layer} object}

\item{...}{Parameters passed to disptach that are used in sorting}
}
\value{
An ordered data.frame if a data.frame was passed. Or nothing in the
  case of a tplyr_layer. These are adjusted silently and will be ordered
  when the table is output.
}
\description{
Named this way to prevent conflicts
}
\examples{
library(dplyr)


#### Sorting output of a table ####
t <- tplyr_table(mtcars, gear) \%>\%
  add_total_group() \%>\%
  add_layer(
    group_count(cyl)
  )
build(t) \%>\%
  tplyr_order(4, 5, Total)
# This orders the column as 4, 5, Total, 3
# The 3 appears at the end because it wasn't passed.

}
