% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{tplyr_table}
\alias{tplyr_table}
\title{Safety Table}
\usage{
tplyr_table(table, headers, header_n)
}
\arguments{
\item{table}{Source data used to create the table. A 'data.frame' object.}

\item{headers}{Column display headers}

\item{header_n}{Population count information to dispaly in headers. A
data.frame with population level information}
}
\value{
A safety_table object which is an environment where the code
  creating the table is evaluated.
}
\description{
The function `tplyr_table()` creates a 'tplyr_table' object which is composed
of a table and table display options, and logic for creating the table. The
layer environments passed to a 'tplyr_table' are used to evaluate the code
used to create the table.
}
\section{Environment Bindings/Properties}{

\itemize{
\item{table - The main piece of data the table will be created from.}
\item{headers - The column headers for the table. These will be repeated
  across pages as appropriate.}
\item{header_n - The population count information used as the denominator
  and display in the column headers.}
\item{layers - The logic used to prepare and render the table.}
}
}

\examples{
df <- data.frame(a = 1:10, b = 11:20)
headers <- c("a", "b")
header_n <- c(a = "10", b = "10")

tplyr_table(df, headers, header_n)
}
\seealso{
[layer()]
}
