% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_bindings.R
\name{tplyr_where}
\alias{tplyr_where}
\title{Set or return where layer binding}
\usage{
tplyr_where(layer)
}
\arguments{
\item{layer}{A \code{tplyr_layer} object.}
}
\value{

}
\description{
Set or return where layer binding
}
