% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_row_masks}
\alias{apply_row_masks}
\title{Replace repeating row label variables with blanks in preparation for display.}
\usage{
apply_row_masks(dat)
}
\arguments{
\item{dat}{Data.frame / tibble to mask repeating row_labels}
}
\value{
tibble with blanked out rows where appropriate
}
\description{
Replace repeating row label variables with blanks in preparation for display.
}
