% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_bindings.R
\name{tplyr_by}
\alias{tplyr_by}
\alias{set_tplyr_by}
\title{Set or return by layer binding}
\usage{
tplyr_by(layer)

set_tplyr_by(layer, by)
}
\arguments{
\item{layer}{A \code{tplyr_layer} object}

\item{by}{A string, a variable name, or a list of variable names supplied
using \code{dplyr::vars}.}
}
\value{
For \code{tplyr_by}, the by binding of the supplied layer. For
  \code{set_tplyr_by} the modified layer environment.
}
\description{
Set or return by layer binding
}
