% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{construct_count_string}
\alias{construct_count_string}
\title{Format n counts for display in count_layer}
\usage{
construct_count_string(
  .n,
  .total,
  .distinct_n = NULL,
  .distinct_total = NULL,
  count_fmt = NULL,
  max_layer_length,
  max_n_width,
  missing_string,
  missing_f_str,
  summary_var
)
}
\arguments{
\item{.n}{Vector of counts for each cell}

\item{.total}{Vector of totals. Should be the same length as .n and be the
denominator that column is based off of.}

\item{.distinct_n}{Vector of distinct counts}

\item{.distinct_total}{Vector of total counts for distinct}

\item{count_fmt}{The f_str object the strings are formatted around.}

\item{max_layer_length}{The maximum layer length of the whole table}

\item{max_n_width}{The maximum length of the actual numeric counts}

\item{missing_string}{The value of the string used to note missing. Usually NA}

\item{missing_f_str}{The f_str object used to display missing values}

\item{summary_var}{The summary_var values that contain the values of the
target variable.}
}
\value{
A tibble replacing the original counts
}
\description{
left padding = (maximum_n_width - this_n_width)
right padding = (maximum_layer_width - this_layer_width[after left padding])
}
