% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numeric.R
\name{get_numeric_data}
\alias{get_numeric_data}
\title{Retrieve the numeric data from a tplyr objects}
\usage{
get_numeric_data(x, ...)
}
\arguments{
\item{x}{A tplyr_table or tplyr_layer object}

\item{...}{Arguments to pass foward}
}
\value{
Numeric data from the Tplyr layer
}
\description{
It may be desirable to access the numeric data from a tplyr object directly, and these objects provide
a method of accessing that directly.
}
\details{
When used on a \code{tplyr_table} object, this method will aggregate the numeric data from all Tplyr layers. The data
are combined into a single data frame and returned to the user. If the data has already been processed (i.e. \code{build} has
been run), the numeric data is already available and will simply be returned. Otherwise, the numeric portion of the layer will
be processed.
}
\examples{
#TBD
}
