% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layering.R
\name{group_count}
\alias{group_count}
\alias{group_desc}
\alias{group_shift}
\title{Create a \code{count}, \code{desc}, or \code{shift} layer for discrete count based summaries, descriptive statistics summaries,
 or shift count summaries}
\usage{
group_count(parent, target_var, by = vars(), where = TRUE, ...)

group_desc(parent, target_var, by = vars(), where = TRUE, ...)

group_shift(parent, target_var, by = vars(), where = TRUE, ...)
}
\arguments{
\item{parent}{\code{tplyr_table} or \code{tplyr_layer}. Required. The parent environment of the layer. This must be either the
\code{tplyr_table} object that the layer is contained within, or another \code{tplyr_layer} object of which
the layer is a subgroup.}

\item{target_var}{Symbol. Required, The variable name on which the summary is to be performed. Must be a variable within
the target dataset. Enter unquoted - i.e. target_var = AEBODSYS.}

\item{by}{A string, a variable name, or a list of variable names supplied using \code{dplyr::vars}}

\item{where}{Call. Filter logic used to subset the target data when performing a summary.}

\item{...}{Additional arguments that will be passed directly into the \code{tplyr_layer} environment. See the
\href{<link tbd>}{vignette} on adding extensions.}
}
\value{
An \code{tplyr_layer} environment that is a child of the specified parent. The environment contains the object
  as listed below.

A \code{tplyr_layer} object
}
\description{
This family of functions specifies the type of summary that is to be performed within a layer. \code{count} layers are used
to create summary counts of some discrete variable. \code{desc} layer create summary statistics, and \code{shift} layers
summaries the counts of different changes in states. See the "details" section below for more information.
}
\details{
#TODO: Complete this section with more information about each layer type.
}
\section{\code{tplyr_layer} Object Structure}{

\describe{
\item{\code{type}}{This is an attribute. A string indicating the layer type, which controls the summary that will be performed.}
\item{\code{target_var}}{A quosure of a name, which is the variable on which a summary will be performed.}
\item{\code{by}}{A list of quosures representing either text labels or variable names used in grouping. Variable names must exist
within the target dataset Text strings submitted do not need to exist in the target dataset.}
\item{\code{where}}{A quosure of a call that containers the filter logic used to subset the target dataset.}
\item{\code{sort_vars}}{A character vector containingn the variables that will be used to sort the results of the summary.
  Set by default to the value of \code{target_var}}
\item{\code{sort}}{A string containing the sort method. Defaults to 'asc' for ascending.}
\item{\code{layers}}{A list with class \code{tplyr_layer_container}. Initialized as empty, but serves as the container for
any sublayers of the current layer.}
\item{\code{formatter}}{A function used to create the string formats for the resulting numbers in output presentation.}
}
}

\examples{
# Load in pipe
library(magrittr)

t <- tplyr_table(iris, Sepal.Width) \%>\%
  add_layer(
    group_desc(target_var=Species)
  )
}
\seealso{
[add_layer(), tplyr_table, tplyr_layer()]

Other Layer construction functions: 
\code{\link{tplyr_layer}()}
}
\concept{Layer construction functions}
