% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_bindings.R
\name{tplyr_pop_data}
\alias{tplyr_pop_data}
\alias{tplyr_pop_data<-}
\alias{set_tplyr_pop_data}
\title{Return or set population data bindings}
\usage{
tplyr_pop_data(table)

tplyr_pop_data(x) <- value

set_tplyr_pop_data(table, pop_data)
}
\arguments{
\item{table}{A \code{tplyr_table} object to set or return population data
information.}

\item{x}{A \code{tplyr_table} object.}

\item{value}{A data.frame with population level information}

\item{pop_data}{A data.frame with population level information}
}
\value{
For \code{tplyr_pop_data} the pop_data binding of the \code{tplyr_table}
  object. For \code{tplyr_pop_data<-} nothing is returned, the pop_data binding
  is set silently. For \code{set_tplyr_pop_data} the modified object.
}
\description{
The population data is used when calculating N counts in the column headers,
as well as determining denominators for percent calculations in the table
body.
}
