% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{process_formatting}
\alias{process_formatting}
\title{Process layers to get formatted and pivoted tables.}
\usage{
process_formatting(x, ...)
}
\arguments{
\item{x}{A tplyr_layer object}

\item{...}{arguments passed to dispatch}
}
\value{
The formatted_table object that is binded to the layer
}
\description{
Process layers to get formatted and pivoted tables.
}
