% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{process_summaries}
\alias{process_summaries}
\title{Process layers to get numeric results of layer}
\usage{
process_summaries(x, ...)
}
\arguments{
\item{x}{a tplyr_layer object}

\item{...}{arguments passed to dispatch}
}
\value{
The tplyr_layer object with a 'built_table' binding
}
\description{
This is an internal method, but is exported to support S3 dispatch. Not intended for direct use by a user.
}
\keyword{internal}
