% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_denom_ignore}
\alias{set_denom_ignore}
\title{Set values the denominator calculation will ignore}
\usage{
set_denom_ignore(e, ...)
}
\arguments{
\item{e}{A count_layer object}

\item{...}{Values to exclude from the percentage calculation.}
}
\value{
The modified layer object
}
\description{
This is generally used for missing values. Values like "", NA, "NA" are
common ways missing values are presented in a data.frame. In certain cases,
percentages do not use "missing" values in the denominator. This function
notes different values as "missing" and excludes them from the denominators.
}
