% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{set_format_strings}
\alias{set_format_strings}
\title{Set the format strings and associated summaries to be performed in a layer}
\usage{
set_format_strings(e, ...)
}
\arguments{
\item{e}{Layer on which to bind format strings}

\item{...}{Named parmeters containing calls to \code{f_str} to set the format strings}
}
\value{
The layer environment with the format string binding added
}
\description{
Set the format strings and associated summaries to be performed in a layer
}
\examples{
TBD
}
