% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bindings.R
\name{set_nest_count}
\alias{set_nest_count}
\title{Set the option to nest count layers}
\usage{
set_nest_count(e, nest_count)
}
\arguments{
\item{e}{A \code{count_layer} object}

\item{nest_count}{A logical value to set the nest option}
}
\value{
The modified layer
}
\description{
If set to TRUE, the second variable specified in \code{target_var}
will be nested inside of the first variable. This allows you to create
displays like those commonly used in adverse event tables, where
one column holds both the labels of the outer categorical variable
and the inside event variable (i.e. AEBODSYS and AEDECOD).
}
