% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_bindings.R
\name{get_sort_vars}
\alias{get_sort_vars}
\alias{set_sort_vars}
\title{Return or set sort_vars layer binding}
\usage{
get_sort_vars(layer)

set_sort_vars(layer, sort_vars)
}
\arguments{
\item{layer}{A \code{tplyr_layer} object}

\item{sort_vars}{A character vector to sort the results of the summary.}
}
\value{
For \code{sort_vars}, the bindings of the layer object. For
  \code{set_sort_vars}, the modified layer environment.
}
\description{
Return or set sort_vars layer binding
}
\examples{
iris$Species2 <- iris$Species
lay <- tplyr_table(iris, Species) \%>\%
  group_count(Species) \%>\%
  set_sort_vars(vars(Sepal.Length, Sepal.Width))
}
