% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{tplyr}
\alias{Tplyr}
\alias{Tplyr-package}
\title{A grammar of summary data for clinical reports}
\description{
\lifecycle{experimental}
}
\details{
'Tplyr' is a package dedicated to simplifying the data manipulation necessary to create clinical reports. Clinical data summaries can
often be broken down into two factors - counting discrete variables (or counting shifts in state), and descriptive statistics around
continuous variable. Many of the reports that go into a clinical report are simply made up of these two scenarios. By abstracting
this process away, 'Tplyr' allows you to rapidly build these tables without worrying about the underlying data manipulation

'Tplyr' takes this process a few steps further by abstracting away all of the programming that goes into proper presentation, which
is where even more of the time is spent when programming these reports. For example, 'Tplyr' allows you to easily control:

\describe{
\item{\strong{Denominators}}{n (\%) counts often vary based on the summary being performed. 'Tplyr' allows you to easily control what denominators
are used based on a few common scenarios, or additionally you can provide a custom dataset used in calculation
\strong{NOTE:} This feature is still being developed and is not yet implemented}
\item{\strong{String formatting}}{Different reports warrant different presentation of your strings. Programming this can get tedious, as you always
want to make sure that your decimals properly align. 'Tplyr' abstracts this process away and provides you with a simple format to specify
how you want your data presented}
\item{\strong{Treatment groups}}{Need a total column? Need to group summaries of multiple treatments? 'Tplyr' makes this simple to add additional
treatment groups into your report}
}

Furthermore, 'Tplyr' was built to be extensible. A common pitfall in creating these solutions is over-automation. Some utilities just do
to much, but don't solve all of the problems. We know we won't handle every scenario - and therefore we've left channels open for you to
add your own touches where necessary.
}
\author{
\strong{Maintainer}: Mike Stackhouse \email{mike.stackhouse@atorusresearch.com} (\href{https://orcid.org/0000-0001-6030-723X}{ORCID})

Authors:
\itemize{
  \item Eli Miller \email{Eli.Miller@AtorusResearch.com} (\href{https://orcid.org/0000-0002-2127-9456}{ORCID})
  \item Ashley Tarasiewicz \email{Ashley.Tarasiewicz@atorusresearch.com}
  \item Nathan Kosiba \email{Nathan.Kosiba@atorusresearch.com} (\href{https://orcid.org/0000-0001-5359-4234}{ORCID})
}

Other contributors:
\itemize{
  \item Atorus Research LLC [copyright holder]
}

}
