% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{tplyr_layer}
\alias{tplyr_layer}
\title{Create a \code{tplyr_layer} object}
\usage{
tplyr_layer(parent, type, by, target_var, where, ...)
}
\arguments{
\item{parent}{\code{tplyr_table} or \code{tplyr_layer}. Required. The parent environment of the layer. This must be either the
\code{tplyr_table} object that the layer is contained within, or another \code{tplyr_layer} object of which
the layer is a subgroup.}

\item{type}{"count", "desc", or "shift". Required. The category of layer - either "counts" for categorical counts, "desc" for
descriptive statistics, or "shift" for shift table counts}

\item{by}{Character Vector. Required. Variable to group the summary being performed, or label of the group to be provided
in the stub. Can be a combination of both.}

\item{target_var}{Symbol. Required, The variable name on which the summary is to be performed. Must be a variable within
the target dataset. Enter unquoted - i.e. target_var = AEBODSYS.}

\item{where}{Call. Filter logic used to subset the target data when performing a summary.}

\item{...}{Additional arguments that will be passed directly into the \code{tplyr_layer} environment. See the
\href{<link tbd>}{vignette} on adding extensions.}
}
\value{
An \code{tplyr_layer} environment that is a child of the specified parent. The environment contains the object
  as listed below.
}
\description{
This object is the workhorse of the \code{tplyr} package. A \code{tplyr_layer} can be thought of
as a block, or "layer" of a table. Summary tables typically consist of different sections that require
different summaries. When programming these section, your code will create different layers that need to be
stacked or merged together. A \code{tplyr_layer} is the container for those isolated building blocks.

When building the \code{tplyr_table}, each layer will execute independently. When all of the data processing has
completed, the layers are brought together to construct the output.

\code{tplyr_layer} objects are not created directly, but rather the layer constructor functions \code{add_group_counts},
\code{add_group_desc}, \code{add_group_shift}, and their related subgroup functions.

See the \href{<link tbd}{vignette} for a more extensive explanation of how these pieces fit together.
}
\section{\code{tplyr_layer} Object Structure}{

}

\examples{
tab <- env(test_data = c(1,2,3,4))
class(tab) <- append('tplyr_table', class(tab))

l <- tplyr_layer(tab, type='count', by="AEBODSYS", target_var=AEDECOD, where= AESER == "Y")

}
\seealso{
\code{\link{tplyr_table}}
}
