% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{tplyr_table}
\alias{tplyr_table}
\title{Safety Table}
\usage{
tplyr_table(target)
}
\arguments{
\item{target}{Source data used to create the table. A 'data.frame' object.}
}
\value{
A safety_table object which is an environment where the code
  creating the table is evaluated.
}
\description{
The function `tplyr_table()` creates a 'tplyr_table' object which is composed
of a table and table display options, and logic for creating the table. The
layer environments passed to a 'tplyr_table' are used to evaluate the code
used to create the table.
}
\section{Environment Bindings/Properties}{

\itemize{
\item{target - The main piece of data the table will be created from.}
\item{headers - The column headers for the table. These will be repeated
  across pages as appropriate.}
\item{pop_data - The data.frame containing population data.}
\item{treat_var - The treatment variable arms/sets are split up on.}
\item{layers - The logic used to prepare and render the table. Stored as a
  list binded to the environment. Use <add layer function> to add a layer.}
}
}

\examples{
df <- data.frame(a = 1:9, b = 10:18, c = rep(c("Placebo", "Low", "High"), 3))
adsl <- data.frame(d = letters[1:9])

tab <- tplyr_table(df) \%>\%
  set_tplyr_treat_var(c) \%>\%
  set_tplyr_pop_data(adsl) \%>\%
  set_tplyr_header(c("Placebo", "Low", "High"))
}
\seealso{
[layer()]
}
