% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{tplyr_table}
\alias{tplyr_table}
\title{Create a Tplyr table object}
\usage{
tplyr_table(target, treat_var, where = TRUE, cols = vars())
}
\arguments{
\item{target}{Dataset upon which summaries will be performed}

\item{treat_var}{Variable containing treatment group assignments. Supply unquoted.}

\item{where}{A general subset to be applied to all layers. Supply as programming logic (i.e. x < 5 & y == 10)}

\item{cols}{A grouping variable to summarize data by column (in addition to treat_var). Provide multiple
column variables by using \code{\link[dplyr]{vars}}}
}
\value{
A \code{tplyr_table} object
}
\description{
The \code{tplyr_table} object is the main container upon which a Tplyr table is constructed. Tplyr tables are made up of
one or more layers. Each layer contains an instruction for a summary to be performed. The \code{tplyr_table} object contains
those layers, and the general data, metadata, and logic necessary.
}
\details{

}
\examples{

tab <- tplyr_table(iris, species, where = Sepal.Length < 5.8)

}
