% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_data.R, R/table_bindings.R
\name{add_total_group}
\alias{add_total_group}
\alias{treat_grps}
\alias{set_treat_grps}
\alias{add_treat_group}
\title{Add a treatment group for the entire dataset population}
\usage{
add_total_group(table)

treat_grps(table)

set_treat_grps(table, group_name, groupings)

add_treat_group(table, group_name, groupings)
}
\arguments{
\item{table}{A tplyr_table object}

\item{group_name}{A character vector with the treatment group names.}

\item{groupings}{A character vector specifiying the treatment variable names.}
}
\value{
The modified table object

For \code{treat_grps} the treat_grp binding of the \code{tplyr_talbe}
  object. For \code{treat_grps<-} and \code{set_treat_grps} the modified
  object. \code{add_treat_group} adds a treatment group without removing others.
}
\description{
This function is a wrapper around \code{add_treat_group} and creates a group
named 'Total' containing the set of unique values in the pop_treat_var
variable.

Treatment groupings are used to create combenations of groups to group. TODO:
Add 'total' option
}
\examples{
tab <- tplyr_table(iris, Species)

add_total_group(tab)
treat_grps(tab)
# Returns "setosa", "verisicolor", and "virginica"

tab <- tplyr_table(iris, Species)

add_treat_group(tab, "Total", as.character(unique(iris$Species)))

}
