% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_data.R, R/table_bindings.R
\name{add_total_group}
\alias{add_total_group}
\alias{treat_grps}
\alias{set_treat_grps}
\alias{add_treat_group}
\title{Add a treatment group for the entire dataset population}
\usage{
add_total_group(table)

treat_grps(table)

set_treat_grps(table, group_name, groupings)

add_treat_group(table, group_name, groupings)
}
\arguments{
\item{table}{A tplyr_table object}

\item{group_name}{A character vector with the treatment group names to be assigned}

\item{groupings}{A character vector specifiying the treatment variable values to be used to construct the assigned group}
}
\value{
The modified table object

For \code{treat_grps} the treat_grp binding of the \code{tplyr_talbe}
  object. For \code{treat_grps<-} and \code{set_treat_grps} the modified
  object. \code{add_treat_group} adds a treatment group without removing others.
}
\description{
This function is a wrapper around \code{add_treat_group} and creates a group
named 'Total' containing the set of unique values in the pop_treat_var
variable.

Treatment groups are used to create additional groups that will be analyzed. For example,
you could create a group of treated subjects vs. placebo in a trial with multiple treatment arms.
These groups are constructed by supplying the assigned name to be used in the data, and then the
which the larger group will be constructed from (i.e. given Placebo, T1, and T2 - Treated would be c('T1', 'T2'))
}
\examples{
tab <- tplyr_table(iris, Species)

add_total_group(tab)
treat_grps(tab)
# Returns "setosa", "verisicolor", and "virginica"

tab <- tplyr_table(iris, Species)

add_treat_group(tab, "v-species", c('versicolor', 'virigina'))

}
